requirements.txt
# BigQuery Stock Quotes Extractor - Dependencies
# Python 3.13 compatible

# Google Cloud BigQuery client
google-cloud-bigquery>=3.25.0

# Environment variable management
python-dotenv>=1.0.0

# Retry/backoff library
tenacity>=8.2.3

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

